pyproject.toml
[project]
name = "okta-mcp-server"
version = "0.1.0"
description = "Okta MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "mcp[cli]>=1.26.0,<2.0.0",
    "okta>=2.9.13",
    "PyJWT[crypto]>=2.4.0",
    "requests>=2.32.4",
    "ruff>=0.11.13",
    "keyring>=25.6.0",
    "keyrings.alt>=5.0.0",
    "flatdict>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/okta_mcp_server"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
okta-mcp-server = "okta_mcp_server:main"
